+++ CHANGELOG.md
@@ -45,6 +45,10 @@ improves testablity and avoids tight coupling. Update your references if applica
 |-|-|-|
 | `Set::hasLoop()` | `Algorithm\Loop::hasLoop()` | [#51](https://github.com/clue/graph/issues/51) |
 | `Vertex::hasLoop()` | `Algorithm\Loop::hasLoopVertex()` | [#51](https://github.com/clue/graph/issues/51) |
+|-|-|-|
+| `Set::hasEdgeParallel()` | `Algorithm\Parallel::hasEdgeParallel()` | [#52](https://github.com/clue/graph/issues/52) |
+| `Edge\Base::hasEdgeParallel()` | `Algorithm\Parallel::hasEdgeParallelEdge()` | [#52](https://github.com/clue/graph/issues/52) |
+| `Edge\Base::getEdgesParallel()` | `Algorithm\Parallel::getEdgeParallelEdge()` | [#52](https://github.com/clue/graph/issues/52) |
 
 * BC break: Remove unneeded algorithm alias definitions ([#31](https://github.com/clue/graph/issues/31), [#50](https://github.com/clue/graph/issues/50)). The following *alias definitions*
 have been removed, their original/actual name has already existed before and continues to work
@@ -76,8 +80,10 @@ unchanged. Update your references if applicable:
     * New `Algorithm\Flow` ([#30](https://github.com/clue/graph/issues/30), [#47](https://github.com/clue/graph/issues/47))
     * New `Algorithm\Tree` ([#44](https://github.com/clue/graph/issues/44))
     * New `Algorithm\Loop` ([#51](https://github.com/clue/graph/issues/51))
+    * New `Algorithm\Parallel` ([#52](https://github.com/clue/graph/issues/52))
 * Feature: `Graph::createVertices()` now also accepts an array of vertex IDs ([#19](https://github.com/clue/graph/issues/19))
 * Fix: Various issues with `Vertex`/`Edge` layout attributes ([#32](https://github.com/clue/graph/issues/32))
+* Fix: Getting multiple parallel edges for undirected edges ([#52](https://github.com/clue/graph/issues/52))
 
 ## 0.5.0 (2013-05-07)